Refactor fakes into a module.
Rename Technology.required -> Technology.cost.
I've realized that my code wrapping technology requirement creation into county creation doesn't let me determine requirements at run time. This is bad and I'll probably have to revert that change.
