+++ undyingkingdoms/models/addons/tech_addons.py
@@ -37,9 +37,10 @@ def available_techs_addon(cls):
 
         Note: for reuse convert to list or set.
         """
+        bp()
         completed_techs = set(self.completed_techs)
         for tech in self.incomplete_techs:
-            if (set(tech.requirements) - completed_techs) is None:
+            if not (set(tech.requirements) - completed_techs):
                 yield tech
 
     cls.available_techs = hybrid_property(get_available_techs)
